+++ R/backend-postgres.R
@@ -248,6 +248,18 @@ sql_translation.PqConnection <- function(con) {
       date_build = function(year, month = 1L, day = 1L, ..., invalid = NULL) {
         sql_expr(make_date(!!year, !!month, !!day))
       },
+      date_count_between = function(start, end, precision, ..., n = 1L){
+
+        check_dots_empty()
+        if (precision != "day") {
+          cli::cli_abort('The only supported value for {.arg precision} on SQL backends is "day"')
+        }
+        if (n != 1) {
+          cli::cli_abort('The only supported value for {.arg n} on SQL backends is "1"')
+        }
+
+        sql_expr(!!end - !!start)
+      },
       get_year = function(x) {
         sql_expr(date_part('year', !!x))
       },